requirements.txt
aiogram~=3.20.0.post0
python-dotenv~=1.1.1
pytz~=2025.2
APScheduler~=3.11.0